requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.9
proxmoxer==2.0.1
requests==2.31.0
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
python-dotenv==1.0.0
websockets==11.0.3
asyncio==3.4.3
pytest==7.4.3
ipaddress==1.0.23
httpx==0.25.0 